pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.120.1",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.2.12",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.27",
]
